refactor(client): move data-lake browse hooks out of fabFiles into dataLakes
