+++ apps/client/app/components/datalake/DataLakeExplorer.nav.test.tsx
@@ -7,7 +7,8 @@ import DataLakeExplorer from './DataLakeExplorer';
 
 // A 2-level tag structure so buildTagTree yields prefix -> children with counts. The Explorer
 // exposes the richest second-level branches (sorted, top 6) + tree navigation via context.
-vi.mock('@client/app/hooks/data/fabFiles', () => ({
+vi.mock('@client/app/hooks/data/dataLakes', () => ({
+  activeOrgId: () => undefined,
   useGetDataLakeTagCounts: () => ({
     data: {
       tagCounts: [
