+++ apps/client/app/hooks/data/fabFiles.ts
@@ -1,10 +1,8 @@
 import {
   CreateFabFileRequestInputType,
-  DATA_LAKES,
   IShareableDocument,
   KnowledgeType,
   UpdateFabFileRequestInputType,
-  normalizeTagPrefix,
   type IFabFileDocument,
 } from '@bike4mind/common';
 import { api } from '@client/app/contexts/ApiContext';
@@ -737,96 +735,6 @@ export function useApplyAutoRenameFabFile() {
   });
 }
 
-export interface DataLakeArticlesParams {
-  id?: string;
-  tags?: string[];
-  search?: string;
-  page?: number;
-  limit?: number;
-  sortBy?: 'fileName' | 'createdAt';
-  sortDir?: 'asc' | 'desc';
-}
-
-/**
- * Fetches tag counts for the Data Lake Explorer tag tree via server-side aggregation.
- * Much lighter than fetching all articles - returns ~50 tag/count pairs instead of 2000 documents.
- */
-export interface DataLakeTagCountsResponse {
-  /** Tag-occurrence sums that drive the Data Lake Explorer's tag tree. */
-  tagCounts: { tag: string; count: number }[];
-  /** Distinct-file counts: combined total + per-prefix breakdown (keyed by lake tag prefix, e.g. 'opti:'). */
-  uniqueArticleCounts: { total: number; byPrefix: Record<string, number> };
-  /**
-   * Distinct live files per lake, keyed by `datalakeTag`. This is the number to show for a
-   * LAKE: it counts membership, so it stays truthful for files that carry no taxonomy tag and
-   * counts a multi-tagged file once. The prefix/occurrence counts above still drive the tag
-   * tree's branches.
-   */
-  lakeFileCounts: Record<string, number>;
-}
-
-/**
- * Which browse surface is reading. Both sources now hit the SAME consolidated
- * `/api/data-lakes/*` endpoints (the former product-gated `/api/opti/*` twins
- * were consolidated away - access is lake-scoped via each lake's declared
- * tag/entitlement gate, so the caller's accessible scope is identical either
- * way). The source is kept as a cache-key discriminator for the two UIs.
- */
-export type DataLakeBrowseSource = 'opti' | 'datalakes';
-const browseBase = (_source: DataLakeBrowseSource) => '/api/data-lakes';
-
-export function useGetDataLakeTagCounts(source: DataLakeBrowseSource = 'opti') {
-  return useQuery({
-    queryKey: ['dataLakeTagCounts', source],
-    queryFn: async () => {
-      const response = await api.get<DataLakeTagCountsResponse>(`${browseBase(source)}/tag-counts`);
-      return response.data;
-    },
-    refetchOnWindowFocus: false,
-    staleTime: 1000 * 60 * 5,
-  });
-}
-
-/**
- * Truthful Data-Lake article counts (distinct files, NOT tag occurrences) for the hero
- * tickers + mission chips, sourced from the same query the Explorer uses. `total` is the
- * combined unique count; the per-prefix fields are the individual per-lake unique counts.
- * Returns 0 for users without data-lake access (the endpoint yields an empty set); callers
- * fall back to a placeholder rather than rendering "0".
- *
- * `sales` is the unique count for the premium (overlay-contributed) lake - its tag prefix is
- * read from the lake config (DATA_LAKES) rather than hardcoded, so no customer-specific prefix
- * lives in open-core; it is 0 in the fork where no premium lake is contributed.
- */
-export function useDataLakeArticleCounts(): { total: number; sales: number; opti: number } {
-  const { data } = useGetDataLakeTagCounts();
-  const unique = data?.uniqueArticleCounts;
-  // The premium lake (if any) is whatever the overlay contributes beyond the base opti lake.
-  const premiumLake = DATA_LAKES.find(l => l.id !== 'opti-knowledge');
-  // `byPrefix` is keyed by the NORMALIZED prefix, and the premium lake's comes from a JSON env
-  // var that is only checked for truthiness - so index it through the same predicate or a
-  // padded value silently reads 0.
-  const premiumPrefix = premiumLake ? normalizeTagPrefix(premiumLake.fileTagPrefix) : null;
-  return {
-    total: unique?.total ?? 0,
-    sales: premiumPrefix ? (unique?.byPrefix[premiumPrefix] ?? 0) : 0,
-    opti: unique?.byPrefix['opti:'] ?? 0,
-  };
-}
-
-export function useGetDataLakeArticles(params?: DataLakeArticlesParams | null, source: DataLakeBrowseSource = 'opti') {
-  return useQuery({
-    queryKey: ['dataLakeArticles', source, params],
-    queryFn: async () => {
-      const response = await api.get<{ data: IFabFileDocument[]; total: number; hasMore: boolean }>(
-        `${browseBase(source)}/articles`,
-        { params: params ?? undefined }
-      );
-      return response.data;
-    },
-    // Disabled when params is null/undefined (lazy-load pattern)
-    enabled: params != null,
-    refetchOnWindowFocus: false,
-    staleTime: 1000 * 60 * 5,
-  });
-}
+// The premium overlay imports this hook from the fabFiles path (pinned ref). Keep the alias
+// until the overlay repoints its imports, then delete it. New code imports from dataLakes.
+export { useDataLakeArticleCounts } from './dataLakes';
